# src/api/openapi.yml
openapi: 3.0.0
info:
  version: 0.1.0
  title: E2B API
servers:
  - url: https://api.e2b.app
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    # AuthProviderBearerAuth / AuthProviderTeamAuth: B before T in the name
    # so Bearer is validated before Team.
    AuthProviderBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: access_token
    AuthProviderTeamAuth:
      type: apiKey
      in: header
      name: X-Team-ID
    # AdminApiKeyAuth / AdminTeamAuth: alphabetical names ensure token
    # validation runs before team context population.
    AdminApiKeyAuth:
      type: apiKey
      in: header
      name: X-Admin-Token
    AdminTeamAuth:
      type: apiKey
      in: header
      name: X-Team-ID
  parameters:
    templateID:
      name: templateID
      in: path
      required: true
      schema:
        type: string
    buildID:
      name: buildID
      in: path
      required: true
      schema:
        type: string
...
  responses:
    "400":
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    "401":
      description: Authentication error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    "404":
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    "500":
      description: Server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    TemplateStep:
      description: Step in the template build process
      required:
        - type
      properties:
        type:
          type: string
          description: Type of the step
        args:
          default: []
          type: array
          description: Arguments for the step
          items:
            type: string
        filesHash:
          type: string
          description: Hash of the files used in the step
        force:
          default: false
          type: boolean
          description: Whether the step should be forced to run regardless of the cache
    TemplateBuildFileUpload:
      description: Upload link for one build context archive, addressed by its files hash
      required:
        - present
      properties:
        present:
          type: boolean
          description: Whether the archive for this hash is already stored
        url:
          type: string
          description: URL the client should PUT the tar archive to
...
    TemplateBuildStartV2:
      type: object
      properties:
        fromImage:
          type: string
          description: Image to use as a base for the template build
        fromTemplate:
          type: string
          description: Template to use as a base for the template build
        force:
          default: false
          type: boolean
          description: Whether the whole build should be forced to run regardless of the cache
        steps:
          default: []
          description: List of steps to execute in the template build
          type: array
          items:
            $ref: "#/components/schemas/TemplateStep"
        startCmd:
          description: Start command to execute in the template after the build
          type: string
        readyCmd:
          description: Ready check command to execute in the template after the build
          type: string
...
    Error:
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int32
          description: Error code
        message:
          type: string
          description: Error
...
paths:
...
  /v2/templates/{templateID}/builds/{buildID}:
    post:
      summary: Start template build (v2)
      description: Start the build
      tags: [templates]
      security:
        - ApiKeyAuth: []
        - AuthProviderBearerAuth: []
          AuthProviderTeamAuth: []
        - AdminApiKeyAuth: []
          AdminTeamAuth: []
      parameters:
        - $ref: "#/components/parameters/templateID"
        - $ref: "#/components/parameters/buildID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TemplateBuildStartV2"
      responses:
        "202":
          description: The build has started
        "400":
          $ref: "#/components/responses/400"
        "401":
          $ref: "#/components/responses/401"
        "404":
          $ref: "#/components/responses/404"
        "500":
          $ref: "#/components/responses/500"
...
  /templates/{templateID}/files/{hash}:
    get:
      summary: Template build file upload link
      description: Get an upload link for a tar archive containing build context files for one COPY step
      tags: [templates]
      security:
        - AccessTokenAuth: []
        - ApiKeyAuth: []
        - AuthProviderBearerAuth: []
          AuthProviderTeamAuth: []
        - AdminApiKeyAuth: []
          AdminTeamAuth: []
      parameters:
        - $ref: "#/components/parameters/templateID"
        - in: path
          name: hash
          required: true
          schema:
            type: string
            description: Hash of the build context files
      responses:
        "201":
          description: Successfully returned the upload link
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TemplateBuildFileUpload"
        "400":
          $ref: "#/components/responses/400"
        "401":
          $ref: "#/components/responses/401"
        "404":
          $ref: "#/components/responses/404"
        "500":
          $ref: "#/components/responses/500"
...